requirements.txt
Django==1.11.6
django-cors-headers==2.1.0
django-debug-toolbar==1.8
django-discover-runner==1.0
django-environ==0.4.4
django-role-permissions==2.1.0
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
funcy==1.10
gunicorn==19.7.1
mysqlclient==1.3.12
PyJWT==1.5.3
pytz==2017.2
six==1.11.0
sqlparse==0.2.4
stringcase==1.2.0
